all functionality complete. CSS added to item overview component.

+++ my-app/src/Components/ItemOverview/itemOverview.js
@@ -23,11 +23,33 @@ function ItemOverview({items, setCartItems, cartItems}) {
     console.log(currentImageIndex);
   }
 
-  function addToCart() {
-    const updatedCartItems = ([...cartItems, selectedItem]);
-    setCartItems(updatedCartItems);
-    localStorage.setItem('cartItems', JSON.stringify(updatedCartItems));
+//   function addToCart() {
+//     const updatedCartItems = ([...cartItems, selectedItem]);
+//     setCartItems(updatedCartItems);
+//     localStorage.setItem('cartItems', JSON.stringify(updatedCartItems));
+//   }
+
+function addToCart() {
+    const existingCartItem = cartItems.find((item) => item._id === selectedItem._id);
+  
+    if (existingCartItem) {
+      const updatedCartItems = cartItems.map((item) => {
+        if (item._id === selectedItem._id) {
+          return { ...item, quantity: item.quantity + 1 };
+        } else {
+          return item;
+        }
+      });
+      setCartItems(updatedCartItems);
+      localStorage.setItem('cartItems', JSON.stringify(updatedCartItems));
+    } else {
+      const updatedCartItem = { ...selectedItem, quantity: 1 };
+      const updatedCartItems = [...cartItems, updatedCartItem];
+      setCartItems(updatedCartItems);
+      localStorage.setItem('cartItems', JSON.stringify(updatedCartItems));
+    }
   }
+
 
   useEffect(() => {
     console.log('Updated cart items:', cartItems);
@@ -40,9 +62,22 @@ return (
         <button onClick={backClick}> &#8249; </button>
         <img className='item-img' src={itemImage[currentImageIndex].url} alt={selectedItem.title}/>
         <button onClick={handleClick}> &#8250; </button> 
+        <div className='item-details'>
         <p className='item-title'>{selectedItem.title}</p>
         <p className='item-price'>£{selectedItem.price}</p>
+        </div>
         <button className='add-to-basket' onClick={addToCart}>Add to basket</button>
+        <div className='image-grid'>
+      {itemImage.map((image, index) => (
+        <img
+          key={index}
+          className={`grid-image ${index === currentImageIndex ? 'selected' : ''}`}
+          src={image.url}
+          alt={selectedItem.title}
+          onClick={() => setCurrentImageIndex(index)}
+        />
+      ))}
+    </div>
     </div>
 )
